move course content click listeners to distinct classes
